improved code usability
-changed how the current scene was being referenced
-made it easier to add buttons, menus, and text to scenes
-changed how colors are set and added the preset "PURPLE"
-added a quit warning to the demo

+++ include/text_object.h
@@ -27,8 +27,8 @@ protected:
     std::vector<std::string> text;
     void* font;
     int font_size;
-    int font_width;
-    int font_height;
+    float font_width;
+    float font_height;
 public:
     int spacing;//the space between each line
     void change_line(unsigned i, std::string l);
